requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
twilio==7.7.1
urllib3==1.26.9
Werkzeug==2.0.3
